Decouple driver from stopwatch
