+++ drivers/embedded/tensil/platform.h
@@ -8,7 +8,6 @@
 #define TENSIL_PLATFORM_DECODER_TIMEOUT 100
 #define TENSIL_PLATFORM_SAMPLE_BLOCK_SIZE 1024
 
-#define TENSIL_PLATFORM_ENABLE_XTIME
 #define TENSIL_PLATFORM_ENABLE_FILE_SYSTEM
 #define TENSIL_PLATFORM_ENABLE_PRINTF
 
@@ -31,7 +30,6 @@
 #define TENSIL_PLATFORM_DECODER_TIMEOUT 100
 #define TENSIL_PLATFORM_SAMPLE_BLOCK_SIZE 1024
 
-#define TENSIL_PLATFORM_ENABLE_XTIME
 // #define TENSIL_PLATFORM_ENABLE_FILE_SYSTEM
 #define TENSIL_PLATFORM_ENABLE_PRINTF
 
@@ -71,8 +69,6 @@ int flash_read(struct flash *flash, uint8_t *buffer, size_t size, size_t *read_s
 
 #define TENSIL_PLATFORM_DRAM_BUFFER_BASE TENSIL_PLATFORM_PROG_BUFFER_HIGH
 #define TENSIL_PLATFORM_DRAM_BUFFER_HIGH 0x8c000000
-
-#define TENSIL_PLATFORM_STOPWATCH_TIMER_DEVICE_ID XPAR_TMRCTR_0_DEVICE_ID
 
 #endif
